Replace all the calls to coroutine methods
When using this library from OpenResty, calls to coroutine.yield raise
this error: `attempt to yield across C-call boundary`.

The reason is documented in the openresty docs:
https://github.com/openresty/lua-nginx-module#lua-coroutine-yieldingresuming

This patch probably has some performance impact. In our particular case,
it does not matter much, because we only use this library for testing.